create a separate library for synq related, so that we don't have circular import issues
